refactor: adding components to own folders
